Fix argument type in ResetLaraFlex()

+++ TR5Main/Game/Lara/lara_helpers.h
@@ -13,6 +13,6 @@ void DoLaraCrawlVault(ITEM_INFO* item, COLL_INFO* coll);
 void DoLaraCrawlToHangSnap(ITEM_INFO* item, COLL_INFO* coll);
 void DoLaraLean(ITEM_INFO* item, COLL_INFO* coll, int maxAngle, short rate);
 void DoLaraCrawlFlex(ITEM_INFO* item, COLL_INFO* coll, short maxAngle, short rate);
-void ResetLaraFlex(ITEM_INFO* item, float rate);
+void ResetLaraFlex(ITEM_INFO* item, short rate);
 void SetLaraFallState(ITEM_INFO* item);
 void SetLaraFallBackState(ITEM_INFO* item);
